refactor: FeatureHolderFileParser
