+++ src/parsing/models/CMakeLists.txt
@@ -1,7 +1,7 @@
 target_sources(dndmanager
     PRIVATE
-    content_file_parser.cpp
     character_file_parser.cpp
+    feature_holder_file_parser.cpp
     character_class_file_parser.cpp
     character_race_file_parser.cpp
     character_subclass_file_parser.cpp
@@ -11,13 +11,11 @@ target_sources(dndmanager
 
 target_sources(dndmanager_tests
     PRIVATE
-    content_file_parser.cpp
     character_file_parser.cpp
+    feature_holder_file_parser.cpp
     character_class_file_parser.cpp
     character_race_file_parser.cpp
     character_subclass_file_parser.cpp
     character_subrace_file_parser.cpp
     spells_file_parser.cpp
 )
-
-add_subdirectory(features)
