+++ src/parsing/models/character_class_file_parser.hpp
@@ -8,11 +8,11 @@
 
 #include "models/character_class.hpp"
 #include "models/features/feature.hpp"
-#include "parsing/models/content_file_parser.hpp"
+#include "parsing/models/feature_holder_file_parser.hpp"
 
 namespace dnd {
 
-class CharacterClassFileParser : public ContentFileParser {
+class CharacterClassFileParser : public FeatureHolderFileParser {
 private:
     std::unordered_map<std::string, std::shared_ptr<const CharacterClass>>& results;
     std::string character_class_name, character_class_hit_dice;
