+++ src/parsing/models/character_file_parser.hpp
@@ -6,7 +6,7 @@
 #include <unordered_map>
 
 #include "models/character.hpp"
-#include "parsing/models/content_file_parser.hpp"
+#include "parsing/content_file_parser.hpp"
 
 namespace dnd {
