+++ src/parsing/models/character_subclass_file_parser.hpp
@@ -8,11 +8,11 @@
 
 #include "models/character_class.hpp"
 #include "models/features/feature.hpp"
-#include "parsing/models/content_file_parser.hpp"
+#include "parsing/models/feature_holder_file_parser.hpp"
 
 namespace dnd {
 
-class CharacterSubclassFileParser : public ContentFileParser {
+class CharacterSubclassFileParser : public FeatureHolderFileParser {
 private:
     std::unordered_map<std::string, std::shared_ptr<const CharacterSubclass>>& results;
     const std::unordered_map<std::string, std::shared_ptr<const CharacterClass>>& classes;
