--- src/parsing/models/features/CMakeLists.txt
@@ -1,9 +1,0 @@
-target_sources(dndmanager
-    PRIVATE
-    feature_parser.cpp
-)
-
-target_sources(dndmanager_tests
-    PRIVATE
-    feature_parser.cpp
-)
